# src/main/resources/docs/openapi_spec.yaml
openapi: 3.0.3
info:
  title: hello-spring API Docs
  description: |-
    hello-spring 프로젝트 API 문서입니다.
    
    정해진 개발 사항에 대해 문서화를 진행하며, 기능 추가 시 마다 문서를 업데이트, GIT을 통해 코드와 함께 형상관리 할 예정입니다.
    
    프로젝트 링크:
    - [hello-spring](https://github.com/hyeok-kong/hello-spring)
  version: 1.0.0
servers:
  - url: https://localhost:8080/api/
...
paths:
...
  /comments/{postId}:
    parameters:
      - in: path
        name: postId
        required: true
        schema:
          type: integer
    get:
      tags:
        - comments
      summary: show comments at post
      description: search comments by postId
      responses:
        '200':
          description: Successful operation
        '404':
          description: Post not found
    post:
      tags:
        - comments
      summary: save comment at post
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
      responses:
        '200':
          description: Successful operation
        '401':
          description: Not authenticated
        '404':
          description: Post not found
  /comments/{commentId}:
    parameters:
      - in: path
        name: commentId
        required: true
        schema:
          type: integer
    patch:
      tags:
        - comments
      summary: update comment
      description: update comment's content
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
      responses:
        '200':
          description: Successful operation
        '401':
          description: Not authenticated
        '403':
          description: Failed to authorize
        '404':
          description: Comment not found
    delete:
      tags:
        - comments
      summary: delete comment
      responses:
        '200':
          description: Successful operation
        '401':
          description: Not authenticated
        '403':
          description: Failed to authorize
        '404':
          description: Comment not found
...